rename controllers to services and router to controller
